Move locking around a source dependency to the entire VCS working directory
